+++ src/app/gamemaster/locations/page.tsx
@@ -1,12 +1,27 @@
 import { useQueryClient } from "@/hooks/useQueryClient";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 import GMLocationsDashboard from "./GMLocationsDashboard";
+import { Suspense } from "react";
+import { DataTableSkeleton } from "@/components/DataTable/data-table-skeleton";
 
 const GMLocationsPage = (): React.ReactElement => {
     const queryClient = useQueryClient();
     return (
         <HydrationBoundary state={dehydrate(queryClient)}>
-            <GMLocationsDashboard />
+            <Suspense fallback={
+                <DataTableSkeleton 
+                                    columnCount={5}
+                                    rowCount={5}
+                                    searchableColumnCount={2}
+                                    filterableColumnCount={2}
+                                    withPagination={false}
+                                    cellWidths={["10rem", "40rem", "12rem", "12rem", "8rem"]}
+                                    shrinkZero
+                                />    
+            }>            
+                <GMLocationsDashboard />
+            </Suspense>
+            
         </HydrationBoundary>
     )
 }
